# src/docs/swagger.yaml
definitions:
...
  github_com_nazarow_golang-clean_api_dto.CityResponse:
    properties:
      country:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CountryResponse"
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.ColorResponse:
    properties:
      hexCode:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.CompanyResponse:
    properties:
      country:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CountryResponse"
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.CountryResponse:
    properties:
      cities:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CityResponse"
        type: array
      companies:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CompanyResponse"
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.CreateColorRequest:
    properties:
      hexCode:
        maxLength: 7
        minLength: 7
        type: string
      name:
        maxLength: 15
        minLength: 3
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.CreateCompanyRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
      - countryId
      - name
    type: object
...
  github_com_nazarow_golang-clean_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description:
          contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_ColorResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.ColorResponse"
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CompanyResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CompanyResponse"
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Filter"
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Sort"
        type: array
    type: object
...
  github_com_nazarow_golang-clean_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.UpdateColorRequest:
    properties:
      hexCode:
        maxLength: 7
        minLength: 7
        type: string
      name:
        maxLength: 15
        minLength: 3
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.UpdateCompanyRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.ResultCode"
      success:
        type: boolean
      validationErrors:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_validations.ValidationError"
        type: array
    type: object
  github_com_nazarow_golang-clean_api_helper.ResultCode:
    enum:
      - 0
      - 40001
      - 40101
      - 40301
      - 40401
      - 42901
      - 42902
      - 50001
      - 50002
    type: integer
    x-enum-varnames:
      - Success
      - ValidationError
      - AuthError
      - ForbiddenError
      - NotFoundError
      - LimiterError
      - OtpLimiterError
      - CustomRecovery
      - InternalError
  github_com_nazarow_golang-clean_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/colors/:
    post:
      consumes:
        - application/json
      description: Create a Color
      parameters:
        - description: Create a Color
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CreateColorRequest"
      responses:
        "201":
          description: Color response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.ColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Create a Color
      tags:
        - Colors
  /v1/colors/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a Color
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Delete a Color
      tags:
        - Colors
    get:
      consumes:
        - application/json
      description: Get a Color
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: Color response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.ColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get a Color
      tags:
        - Colors
    put:
      consumes:
        - application/json
      description: Update a Color
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
        - description: Update a Color
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.UpdateColorRequest"
      responses:
        "200":
          description: Color response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.ColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Update a Color
      tags:
        - Colors
  /v1/colors/get-by-filter:
    post:
      consumes:
        - application/json
      description: Get Colors
      parameters:
        - description: Request
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter"
      responses:
        "200":
          description: Color response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_ColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get Colors
      tags:
        - Colors
  /v1/companies/:
    post:
      consumes:
        - application/json
      description: Create a Company
      parameters:
        - description: Create a Company
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CreateCompanyRequest"
      responses:
        "201":
          description: Company response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CompanyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Create a Company
      tags:
        - Companies
  /v1/companies/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a Company
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Delete a Company
      tags:
        - Companies
    get:
      consumes:
        - application/json
      description: Get a Company
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: Company response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CompanyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get a Company
      tags:
        - Companies
    put:
      consumes:
        - application/json
      description: Update a Company
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
        - description: Update a Company
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.UpdateCompanyRequest"
      responses:
        "200":
          description: Company response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CompanyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Update a Company
      tags:
        - Companies
  /v1/companies/get-by-filter:
    post:
      consumes:
        - application/json
      description: Get Companies
      parameters:
        - description: Request
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter"
      responses:
        "200":
          description: Company response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CompanyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get Companies
      tags:
        - Companies
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"